+++ src/Illuminate/Notifications/Channels/DatabaseChannel.php
@@ -3,29 +3,9 @@
 namespace Illuminate\Notifications\Channels;
 
 use Illuminate\Notifications\Notification;
-use Illuminate\Contracts\Events\Dispatcher;
-use Illuminate\Notifications\Events\DatabaseNotificationCreated;
 
 class DatabaseChannel
 {
-    /**
-     * The event dispatcher.
-     *
-     * @var \Illuminate\Contracts\Events\Dispatcher
-     */
-    protected $events;
-
-    /**
-     * Create a new database channel.
-     *
-     * @param  \Illuminate\Contracts\Events\Dispatcher  $events
-     * @return void
-     */
-    public function __construct(Dispatcher $events)
-    {
-        $this->events = $events;
-    }
-
     /**
      * Send the given notification.
      *
@@ -36,18 +16,27 @@ class DatabaseChannel
     public function send($notifiables, Notification $notification)
     {
         foreach ($notifiables as $notifiable) {
-            $databaseNotification = $notifiable->routeNotificationFor('database')->create([
-                'level' => $notification->level,
-                'intro' => $notification->introLines,
-                'outro' => $notification->outroLines,
-                'action_text' => $notification->actionText,
-                'action_url' => $notification->actionUrl,
-                'read' => false,
-            ]);
-
-            $this->events->fire(new DatabaseNotificationCreated(
-                $notifiable, $notification, $databaseNotification
-            ));
+            $this->createNotification($notifiable, $notification);
         }
+    }
+
+    /**
+     * Create a database notification record for the notifiable.
+     *
+     * @param  mixed  $notifiable
+     * @param  \Illuminate\Notifications\Notification  $notification
+     * @return \Illuminate\Notifications\DatabaseNotification
+     */
+    protected function createNotification($notifiable, Notification $notification)
+    {
+        return $notifiable->routeNotificationFor('database')->create([
+            'type' => get_class($notification),
+            'level' => $notification->level,
+            'intro' => $notification->introLines,
+            'outro' => $notification->outroLines,
+            'action_text' => $notification->actionText,
+            'action_url' => $notification->actionUrl,
+            'read' => false,
+        ]);
     }
 }
