+++ tests/Notifications/NotificationBroadcastChannelTest.php
@@ -1,16 +1,16 @@
 <?php
 
 use Illuminate\Notifications\Notification;
-use Illuminate\Notifications\Channels\DatabaseChannel;
+use Illuminate\Notifications\Channels\BroadcastChannel;
 
-class NotificationDatabaseChannelTest extends PHPUnit_Framework_TestCase
+class NotificationBroadcastChannelTest extends PHPUnit_Framework_TestCase
 {
     public function tearDown()
     {
         Mockery::close();
     }
 
-    public function testDatabaseChannelCreatesDatabaseRecordWithProperData()
+    public function testBroadcastChannelCreatesDatabaseRecordWithProperData()
     {
         $notification = new Notification;
         $notifiables = collect([$notifiable = Mockery::mock()]);
@@ -25,7 +25,10 @@ class NotificationDatabaseChannelTest extends PHPUnit_Framework_TestCase
             'read' => false,
         ]);
 
-        $channel = new DatabaseChannel;
+        $events = Mockery::mock('Illuminate\Contracts\Events\Dispatcher');
+        $events->shouldReceive('fire')->once()->with('Illuminate\Notifications\Events\DatabaseNotificationCreated');
+
+        $channel = new BroadcastChannel($events);
         $channel->send($notifiables, $notification);
     }
 }
